JBPM-7996 - Externalize data source auto configuration to allow easier migration to latest SpringBoot
