+++ kie-spring-boot/kie-spring-boot-autoconfiguration/jbpm-spring-boot-data-sources/.gitignore
@@ -1,0 +1,19 @@
+/target
+/local
+*.tlog
+
+# Eclipse, Netbeans and IntelliJ files
+/.*
+!.gitignore
+/nbproject
+/*.ipr
+/*.iws
+*.iml
+.classpath
+.project
+.settings
+.idea
+target
+
+# Repository wide ignore mac DS_Store files
+.DS_Store
