+++ kie-spring-boot/kie-spring-boot-autoconfiguration/pom.xml
@@ -12,11 +12,12 @@
 
   <packaging>pom</packaging>
   <modules>
+    <module>jbpm-spring-boot-data-sources</module>
     <module>jbpm-spring-boot-autoconfiguration</module>
     <module>kie-server-spring-boot-autoconfiguration</module>
     <module>kie-server-spring-boot-autoconfiguration-drools</module>
     <module>kie-server-spring-boot-autoconfiguration-jbpm</module>
-    <module>kie-server-spring-boot-autoconfiguration-optaplanner</module>
+    <module>kie-server-spring-boot-autoconfiguration-optaplanner</module>    
   </modules>
 
 </project>
